legacy/requirements.txt
google-generativeai>=0.8.0
google-auth-oauthlib>=1.2.0
google-api-python-client>=2.100.0
